requirements.txt
requests  
langchain-core
python-dotenv
fastapi
uvicorn
langchain 
langchain-chroma
langchain_ollama
PyPDF2
langchain-huggingface